Use TypeVar's bound instead of constraint
Using a factory in mypy strict requires a typed generic and providing
the model itself as the generic type did not work because contrained
TypeVar doesn't allow subclassing, as a bound TypeVar allows it.

A file with this typing with its own mypy config was added to test the
behaviour remains.   Adding this showed that the DataclassProtocol
was too narrow as a single field dataclass does not have
`__dataclass_params__` and `__post_init__` attributes and it seems
that marking them as Optional is not enough.

Fixes #21