Fix: Random union types for dict (#200)
* Random union types for dict

Previously under pydantic-factory
values for a dictionary of union
were generated randomly based on
union types. Under the latest
version of polyfactory always the
first type is used. After that fix
everything is working like previously.

Test for that issue was added.

* Random union types for dict

Previously under pydantic-factory
values for a dictionary of union
were generated randomly based on
union types. Under the latest
version of polyfactory always the
first type is used. After that fix
everything is working like previously.

Test for that issue was added.

* Random union types for dict

Previously under pydantic-factory
values for a dictionary of union
were generated randomly based on
union types. Under the latest
version of polyfactory always the
first type is used. After that fix
everything is working like previously.

Test for that issue was added.

* Random union types for dict

Previously under pydantic-factory
values for a dictionary of union
were generated randomly based on
union types. Under the latest
version of polyfactory always the
first type is used. After that fix
everything is working like previously.

Test for that issue was added.

* Random union types for dict

1. Better approach for testing
provided bug / improvement.