test: remove directories to test from pytest addopts
The 'tests docs/examples' were removed from addopts and instead added to the CI and the Makefile explicitly. This allows for running pytest only on specific files ex: pytest tests/test_msgspec_factory.py to work as expected. Before, all the tests would run because of the addopts, but now only the specified test would run.